requirements.txt
python-telegram-bot==20.3
flask
pymongo
python-dotenv
requests